Revise inernal references in pkg components
The location of these functions caused a cross internal/pkg
reference, which made external SDK builds difficult. This removes
the internal components from the pkg components to ensure
proper pkg compatibility.
