+++ cmd/pgo/cmd/user.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/crunchydata/postgres-operator/cmd/pgo/api"
 	"github.com/crunchydata/postgres-operator/cmd/pgo/util"
+	"github.com/crunchydata/postgres-operator/internal/apiserver"
 	utiloperator "github.com/crunchydata/postgres-operator/internal/util"
 	msgs "github.com/crunchydata/postgres-operator/pkg/apiservermsgs"
 
@@ -89,7 +90,7 @@ func createUser(args []string, ns string) {
 	}
 
 	// determine if the user provies a valid password type
-	if _, err := msgs.GetPasswordType(PasswordType); err != nil {
+	if _, err := apiserver.GetPasswordType(PasswordType); err != nil {
 		fmt.Println("Error:", err.Error())
 		os.Exit(1)
 	}
@@ -400,7 +401,7 @@ func updateUser(clusterNames []string, namespace string) {
 	}
 
 	// determine if the user provies a valid password type
-	if _, err := msgs.GetPasswordType(PasswordType); err != nil {
+	if _, err := apiserver.GetPasswordType(PasswordType); err != nil {
 		fmt.Println("Error:", err.Error())
 		os.Exit(1)
 	}
