+++ pkg/apiservermsgs/usermsgs.go
@@ -15,12 +15,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import (
-	"errors"
-
-	pgpassword "github.com/crunchydata/postgres-operator/internal/postgres/password"
-)
-
 type UpdateClusterLoginState int
 
 // set the different values around whether or not to disable/enable a user's
@@ -30,19 +24,6 @@ const (
 	UpdateUserLoginEnable
 	UpdateUserLoginDisable
 )
-
-// ErrPasswordTypeInvalid is used when a string that's not included in
-// PasswordTypeStrings is used
-var ErrPasswordTypeInvalid = errors.New("invalid password type. choices are (md5, scram-sha-256)")
-
-// passwordTypeStrings is a mapping of strings of password types to their
-// corresponding value of the structured password type
-var passwordTypeStrings = map[string]pgpassword.PasswordType{
-	"":              pgpassword.MD5,
-	"md5":           pgpassword.MD5,
-	"scram":         pgpassword.SCRAM,
-	"scram-sha-256": pgpassword.SCRAM,
-}
 
 // CreateUserRequest contains the parameters that are passed in when an Operator
 // user requests to create a new PostgreSQL user
@@ -152,16 +133,4 @@ type UserResponseDetail struct {
 	Password     string
 	Username     string
 	ValidUntil   string
-}
-
-// GetPasswordType returns the enumerated password type based on the string, and
-// an error if it cannot match one
-func GetPasswordType(passwordTypeStr string) (pgpassword.PasswordType, error) {
-	passwordType, ok := passwordTypeStrings[passwordTypeStr]
-
-	if !ok {
-		return passwordType, ErrPasswordTypeInvalid
-	}
-
-	return passwordType, nil
 }
