+++ pkg/events/eventing.go
@@ -23,15 +23,12 @@ import (
 	"reflect"
 	"time"
 
-	crunchylog "github.com/crunchydata/postgres-operator/internal/logging"
 	"github.com/nsqio/go-nsq"
 	log "github.com/sirupsen/logrus"
 )
 
 // String returns the string form for a given LogLevel
 func Publish(e EventInterface) error {
-	// Add logging configuration
-	crunchylog.CrunchyLogger(crunchylog.SetParameters())
 	eventAddr := os.Getenv("EVENT_ADDR")
 	if eventAddr == "" {
 		return errors.New("EVENT_ADDR not set")
